+++ src/utils.jl
@@ -1,12 +1,12 @@
 """
-    build_components(X::Matrix{Tl}, coefs::Vector{Float64}, components_indexes::Dict{String, Vector{Int64}}) -> Dict where Tl
+    build_components(X::Matrix{Tl}, coefs::Vector{Float64}, components_indexes::Dict{String, Vector{Int}}) -> Dict where Tl
     
     Constructs components dict containing values, indexes and coefficients for each component.
 
     # Arguments
     - X::Matrix{Tl}: Input matrix.
     - coefs::Vector{Float64}: Coefficients.
-    - components_indexes::Dict{String, Vector{Int64}}: Dictionary mapping component names to their indexes.
+    - components_indexes::Dict{String, Vector{Int}}: Dictionary mapping component names to their indexes.
 
     # Returns
     - components::Dict: Dictionary containing components, each represented by a dictionary with keys:
@@ -15,7 +15,7 @@
         - "Values": Values computed from `X` and component coefficients.
 
 """
-function build_components(X::Matrix{Tl}, coefs::Vector{Float64}, components_indexes::Dict{String, Vector{Int64}})::Dict where Tl
+function build_components(X::Matrix{Tl}, coefs::Vector{Float64}, components_indexes::Dict{String, Vector{Int}})::Dict where Tl
     components = Dict()
     for key in keys(components_indexes)
         components[key] = Dict()
@@ -30,7 +30,7 @@ function build_components(X::Matrix{Tl}, coefs::Vector{Float64}, components_inde
 end
 
 """
-get_fit_and_residuals(estimation_ε::Vector{Float64}, coefs::Vector{Float64}, X::Matrix{Tl}, valid_indexes::Vector{Int64}, T::Int64) -> Tuple{Vector{Float64}, Vector{Float64}} where Tl
+get_fit_and_residuals(estimation_ε::Vector{Float64}, coefs::Vector{Float64}, X::Matrix{Tl}, valid_indexes::Vector{Int}, T::Int) -> Tuple{Vector{Float64}, Vector{Float64}} where Tl
 
     Builds complete residuals and fit in sample. Residuals will contain nan values for non valid indexes. Fit in Sample will be a vector of fitted values computed from input data and coefficients (non valid indexes will also be calculated via interpolation).
 
@@ -38,8 +38,8 @@ get_fit_and_residuals(estimation_ε::Vector{Float64}, coefs::Vector{Float64}, X:
     - `estimation_ε::Vector{Float64}`: Vector of estimation errors.
     - `coefs::Vector{Float64}`: Coefficients.
     - `X::Matrix{Tl}`: Input matrix.
-    - `valid_indexes::Vector{Int64}`: Valid indexes.
-    - `T::Int64`: Length of the original time series.
+    - `valid_indexes::Vector{Int}`: Valid indexes.
+    - `T::Int`: Length of the original time series.
 
     # Returns
     - Tuple containing:
@@ -47,40 +47,40 @@ get_fit_and_residuals(estimation_ε::Vector{Float64}, coefs::Vector{Float64}, X:
         - `fitted::Vector{Float64}`: Vector of fitted values computed from input data and coefficients.
 
 """
-function get_fit_and_residuals(estimation_ε::Vector{Float64}, coefs::Vector{Float64}, X::Matrix{Tl}, valid_indexes::Vector{Int64}, T::Int64)::Tuple{Vector{Float64}, Vector{Float64}} where Tl
+function get_fit_and_residuals(estimation_ε::Vector{Float64}, coefs::Vector{Float64}, X::Matrix{Tl}, valid_indexes::Vector{Int}, T::Int)::Tuple{Vector{Float64}, Vector{Float64}} where Tl
     ε      = fill(NaN, T); ε[valid_indexes] = estimation_ε
     fitted = X*coefs
     return ε, fitted
 end
 
 """
-o_size(T::Int64)::Int64
+o_size(T::Int)::Int
 
     Calculates the size of outlier matrix based on the input T.
 
     # Arguments
-    - `T::Int64`: Length of the original time series.
+    - `T::Int`: Length of the original time series.
 
     # Returns
-    - `Int64`: Size of o calculated from T.
+    - `Int`: Size of o calculated from T.
 
 """
-o_size(T::Int64)::Int64 = T
+o_size(T::Int)::Int = T
 
 """
-create_o_matrix(T::Int64, steps_ahead::Int64)::Matrix
+create_o_matrix(T::Int, steps_ahead::Int)::Matrix
 
     Creates a matrix of outliers based on the input sizes, and the desired steps ahead (this is necessary for the forecast function).
 
     # Arguments
-    - `T::Int64`: Length of the original time series.
-    - `steps_ahead::Int64`: Number of steps ahead (for estimation purposes this should be set at 0).
+    - `T::Int`: Length of the original time series.
+    - `steps_ahead::Int`: Number of steps ahead (for estimation purposes this should be set at 0).
     
     # Returns
     - `Matrix`: Matrix of outliers constructed based on the input sizes.
 
 """
-function create_o_matrix(T::Int64, steps_ahead::Int64)::Matrix
+function create_o_matrix(T::Int, steps_ahead::Int)::Matrix
     return vcat(Matrix(1.0 * I, T, T), zeros(steps_ahead, T))
 end
 
@@ -97,9 +97,9 @@ handle_missing_values(X::Matrix{Tl}, y::Vector{Fl}) -> Tuple{Vector{Fl}, Matrix{
     - Tuple containing:
         - `y::Vector{Fl}`: Time series without missing values.
         - `X::Matrix{Tl}`: Input matrix without missing values.
-        - `valid_indexes::Vector{Int64}`: Vector containing valid indexes of the time series.
+        - `valid_indexes::Vector{Int}`: Vector containing valid indexes of the time series.
 """
-function handle_missing_values(X::Matrix{Tl}, y::Vector{Fl})::Tuple{Vector{Fl}, Matrix{Tl}, Vector{Int64}} where {Tl, Fl}
+function handle_missing_values(X::Matrix{Tl}, y::Vector{Fl})::Tuple{Vector{Fl}, Matrix{Tl}, Vector{Int}} where {Tl, Fl}
 
     invalid_indexes = unique(vcat([i[1] for i in findall(i -> any(isnan, i), X)], findall(i -> isnan(i), y)))
     valid_indexes   = setdiff(1:length(y), invalid_indexes)
@@ -134,7 +134,7 @@ fill_innovation_coefs(component::String, output::Output)::Vector{Float64}
     # Returns
     - `Vector{Float64}`: Vector containing innovation coefficients for the given component.
 """
-function fill_innovation_coefs(T::Int64, component::String, output::Output)::Vector{Float64}
+function fill_innovation_coefs(T::Int, component::String, output::Output)::Vector{Float64}
     inov_comp = zeros(T)
     for (i, idx) in enumerate(output.components[component]["Indexes"])
         inov_comp[findfirst(i -> i != 0, output.X[:, idx])] = output.components[component]["Coefs"][i]
